Cleaned up lots of code. ES6 style import useage across the board.
