+++ .babelrc
@@ -1,3 +1,4 @@
 {
-    presets: ['react']
+    'presets': ['react-plus'],
+    'plugins': ['transform-decorators-legacy']
 }
