+++ package.json
@@ -17,27 +17,18 @@
   },
   "devDependencies": {
     "babel-core": "^6.26.0",
+    "babel-eslint": "^8.2.1",
     "babel-loader": "^7.1.2",
+    "babel-plugin-transform-decorators-legacy": "^1.3.4",
+    "babel-preset-mobx": "^1.0.2",
     "babel-preset-react": "^6.24.1",
-    "chai": "^4.1.2",
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^4.1.0",
-    "gulp-concat": "^2.6.1",
-    "gulp-cssnano": "^2.1.2",
-    "gulp-html5-lint": "^1.1.0",
-    "gulp-jshint": "^2.1.0",
-    "gulp-livereload": "^3.8.1",
-    "gulp-sass": "^3.1.0",
-    "gulp-sourcemaps": "^2.6.4",
-    "gulp-uglify": "^3.0.0",
+    "babel-preset-react-plus": "^1.3.0",
+    "clean-webpack-plugin": "^0.1.18",
+    "css-loader": "^0.28.9",
     "html-webpack-plugin": "^2.30.1",
-    "jshint": "^2.9.5",
-    "jshint-stylish": "^2.2.1",
-    "mocha": "^5.0.0",
     "node-sass": "^4.7.2",
-    "pump": "^2.0.1",
-    "uglify-es": "^3.3.9",
-    "uglify-js": "^3.3.9",
+    "sass-loader": "^6.0.6",
+    "style-loader": "^0.20.2",
     "webpack": "^3.11.0",
     "webpack-dev-server": "^2.11.1"
   },
