+++ src/components/SpoopyGUI.jsx
@@ -1,9 +1,11 @@
 import React, {Component} from 'react';
+import {observer} from 'mobx-react';
 
+@observer
 export default class SpoopyGUI extends Component {
     render() {
         return (
-            <h1>SpoopyGUI placeholder element.</h1>
+            <h1>{this.props.player.loc}</h1>
         );
     }
 }
