--- src/react/spoopy.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import SpoopyGUI from './components/SpoopyGUI.jsx';
-
-ReactDOM.render(
-    <SpoopyGUI />,
-    document.getElementById('SpoopyGUI')
-);
